requirements.txt
# Core HTTP/API
requests>=2.31.0
urllib3>=2.0.0

# Configuration & typing
pydantic>=2.5.0
pydantic-settings>=2.1.0
typing-extensions>=4.8.0

# Image processing
Pillow>=10.1.0

# Font handling
fonttools>=4.47.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.1.0
mypy>=1.7.0

# CLI convenience
click>=8.1.0

# Web demo (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
